Make Mouseinput.MouseData unsigned. According to the documentation it should be a dword (uint) but it is impossible to send a negative/scrolldown mouse event this way so it was changed to unsigned.

+++ MetroController/WindowsInput/InputBuilder.cs
@@ -211,7 +211,7 @@ namespace MetroController.WindowsInput {
         {
             var scroll = new Input { Type = (uint) InputType.Mouse };
             scroll.Data.Mouse.Flags = (uint) MouseFlag.VerticalWheel;
-            scroll.Data.Mouse.MouseData = (uint) scrollAmount;
+            scroll.Data.Mouse.MouseData = scrollAmount;
 
             _inputList.Add(scroll);
